attempt to add Herk/Syrk impls to rocBLAS
This also includes a bunch of other updates to the rocBLAS API style
(rocblas_int instead of int, etc).

As a caveat: I can't test this right now since ROCm 3.6.0 seems to
have changed things in the CMake structure and I don't have time to
dig into that in great depth at the moment.