Add Aluminum dispatch for send and recv within TranslateBetweenGrids (#178)
* Add Aluminum dispatch for send and recv within TranslateBetweenGrids

This commit only addresses the synchronous calls in the (STAR,VC) variant. I'm
using this as a bit of testbed before wholesale adding these into regular
`El::mpi` dispatch as the point-to-point use-cases are a bit more delicate.

* Appease linguistic nits being picked offline

* Fix formatting

* Squash build warnings

* Aluminum-ize El::mpi::Send and El::mpi::Recv; Aluminum now required

Note that the nonblocking Send/Recv are still not aluminum-ized. (They also
don't check host/device-ness of buffers, so there's that...) It would be a
bigger lift to incorporate Aluminum with the Request struct, so I'm disinclined
to do that until there's programmatic need.

* Revert to now-Aluminum-ized El::mpi::Send and El::mpi::Recv; cleanup syncs